// verkos-exhibit-src/src/libs/shared/types/hardware.enum.ts
export enum SupportedDrones {
  M30_SERIES = 'M30Series',
  M3D_SERIES = 'M3DSeries',
  M4D_SERIES = 'M4DSeries',
  M3E_SERIES = 'M3ESeries',
  M300_SERIES = 'M300Series',
  M350_SERIES = 'M350Series',
}

export enum SupportedDockingStations {
  DOCK1 = 'Dockv1',
  DOCK2 = 'Dock2v1',
  DOCK3 = 'Dock3v1',
  LEGACY = 'LegacyDock',
}

export enum DronePayloadTypes {
  M30 = 'M30',
  M30T = 'M30T',
  M3D = 'M3D',
  M3TD = 'M3TD',
  M4D = 'M4D',
  M4TD = 'M4TD',
  M3E = 'M3E',
  M3T = 'M3T',
  M300 = 'M300',
  M350 = 'M350',
}